word2vec

+++ tasks/word2vec/word2vec.cpp
@@ -5,11 +5,11 @@
 std::vector<std::string> FindClosestWords(const std::vector<std::string>& words,
                                           const std::vector<std::vector<int>>& vectors) {
     std::vector<std::string> ans;
-    size_t max_len = 0;
-    size_t len = 0;
-    for (int i = 1; i < vectors.size(); ++i) {
+    int64_t max_len = 0;
+    int64_t len = 0;
+    for (size_t i = 1; i < vectors.size(); ++i) {
         for (int coord : vectors[i]) {
-            len += coord * vectors[0][i];
+            len += static_cast<int64_t>(coord) * static_cast<int64_t>(vectors[0][i]);
         }
         if (len == max_len) {
             ans.push_back(words[i]);
